refactor(vision): share narrow read model with Todo projection
